requirements.txt
annoy==1.17.2
clip_interrogator==0.6.0
cog==0.7.0
coloredlogs==15.0.1
decord==0.6.0
easynmt==2.0.2
elasticsearch==8.7.0
fairscale==0.4.4
faiss==1.5.3
hnswlib==0.7.0
huggingface_hub==0.13.3
matplotlib==3.7.1
nlpaug==1.1.11
nltk==3.8.1
numpy==1.24.2
open_clip_torch==2.18.0
opustools==1.5.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pycocoevalcap==1.2
pycocotools==2.0.6
pytorch_lightning==1.8.0
pytrec_eval==0.5
recommonmark==0.7.1
Requests==2.30.0
ruamel_yaml==0.17.22
scikit_learn==1.2.2
scipy==1.10.1
setuptools==44.0.0
Sphinx==7.0.0
timm==0.4.12
torch==1.13.0
torch_xla==1.0
torchmetrics==0.11.4
torchvision==0.14.0
tqdm==4.65.0
transformers==4.15.0
